+++ test/cbmc/proofs/parsing/ProcessReceivedTCPPacket/Makefile.json
@@ -10,7 +10,11 @@
   [
     "$(ENTRY)_harness.goto",
     "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_IP.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_IP.goto"
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_IP.goto",
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_Transmission.goto",
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling.goto",
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_Utils.goto",
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_Reception.goto"
   ],
   "INSTFLAGS":
   [
